requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-socketio==5.9.0
python-socketio[asgi]
python-dotenv==1.0.0 